+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `composer.lock` as it is needed by some services
+- new method `YoutubeDownloader\Provider\Youtube\SignatureDecipher::extractDecipherOpcode()` for extracting the decipher operation codes
+
+### Deprecated
+
+- `YoutubeDownloader\Provider\Youtube\SignatureDecipher::decipherSignatureWithRawPlayerScript()` isn't used anymore and will be removed in 0.8
+
 ### Removed
 
 - **Breaking:** `YoutubeDownloader\Provider\Youtube\Provider::createFromConfigAndToolkit()` was removed, use `YoutubeDownloader\Provider\Youtube\Provider::createFromOptions()` instead